Spawn piece above the board so it falls in.
At the moment the noclip gets turned off immediately after the first
collision so it can't spawn too high or it'll never fall into the board.

What I need is a function to test if the piece is inside the game board.
This will help to prevent a fast falling first piece from falling out
(see first-piece-falls-out branch), and will also make this more robust
so the piece can spawn as high as I want.